+++ js/renovation/ui/scheduler/workspaces/base/group_panel/vertical/row.tsx
@@ -2,8 +2,7 @@ import {
   Component, ComponentBindings, JSXComponent, OneWay, Template,
 } from 'devextreme-generator/component_declaration/common';
 import { GroupPanelVerticalCell } from './cell';
-// eslint-disable-next-line import/extensions
-import { GroupRenderItem } from '../../../types';
+import { GroupRenderItem } from '../../../types.d';
 
 export const viewFunction = (viewModel: Row) => (
   <div
