+++ js/renovation/ui/scheduler/workspaces/base/header_panel/layout.tsx
@@ -2,8 +2,7 @@ import {
   Component, ComponentBindings, JSXComponent, OneWay, Template,
 } from 'devextreme-generator/component_declaration/common';
 import { Row } from '../row';
-// eslint-disable-next-line import/extensions
-import { ViewCellData } from '../../types';
+import { ViewCellData } from '../../types.d';
 import { getKeyByDateAndGroup } from '../../utils';
 
 export const viewFunction = (viewModel: HeaderPanelLayout) => (
