+++ js/renovation/ui/scheduler/workspaces/base/layout_props.ts
@@ -1,8 +1,7 @@
 import {
   ComponentBindings, OneWay,
 } from 'devextreme-generator/component_declaration/common';
-// eslint-disable-next-line import/extensions
-import { GroupedViewData } from '../types';
+import { GroupedViewData } from '../types.d';
 
 @ComponentBindings()
 export class LayoutProps {
