+++ js/renovation/ui/scheduler/workspaces/base/time_panel/layout.tsx
@@ -3,8 +3,7 @@ import {
 } from 'devextreme-generator/component_declaration/common';
 import { Row } from '../row';
 import { TimePanelCell as Cell } from './cell';
-// eslint-disable-next-line import/extensions
-import { ViewCellData } from '../../types';
+import { ViewCellData } from '../../types.d';
 import { getKeyByDateAndGroup } from '../../utils';
 import { Table } from '../table';
 import { LayoutProps } from '../layout_props';
